Change rtw_tx_pwr_idx_override type to s16

+++ hal/rtl8812a/rtl8812a_phycfg.c
@@ -39,7 +39,7 @@ int transmit_power_boost = 10;
 // Transmit Power Override
 // This value completely overrides the driver's calculations and uses only one value for all transmissions.
 // Zero disables it. Default: 0
-extern int rtw_tx_pwr_idx_override;
+extern s16 rtw_tx_pwr_idx_override;
 
 /* Manual Transmit Power Control */
